DjangoAPI/requirements_python.txt
asgiref==3.3.4
boto3==1.17.56
botocore==1.20.56
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
future==0.18.2
jmespath==0.10.0
microscoper==0.3.1
numpy==1.20.2
opencv-python==4.5.1.48
PyJWT==1.7.1
python-bioformats==4.0.4
python-dateutil==2.8.1
python-javabridge==4.0.3
pytz==2021.1
s3transfer==0.4.2
six==1.15.0
sqlparse==0.4.1
tifffile==2021.4.8
tqdm==4.60.0
urllib3==1.26.4
